+++ conversejs/build-conversejs.sh
@@ -18,22 +18,14 @@ set -x
 CONVERSE_VERSION="v11.0.0"
 CONVERSE_REPO="https://github.com/conversejs/converse.js.git"
 # You can eventually set CONVERSE_COMMIT to a specific commit ID, if you want to apply some patches.
-# 2024-08-05: using Converse upstream (v11 WIP).
-CONVERSE_COMMIT="5017efb780973d704f237c478ba52b23d901e1bf"
+# 2024-09-02: using Converse upstream (v11 WIP).
+CONVERSE_COMMIT="9952046d580bc2930e29833f4c9987a3d4c95bc2"
 
 # It is possible to use another repository, if we want some customization that are not upstream (yet):
 # CONVERSE_VERSION="livechat"
 # # CONVERSE_COMMIT="4402fcc3fc60f6c9334f86528c33a0b463371d12"
 # CONVERSE_REPO="https://github.com/JohnXLivingston/converse.js"
 # CONVERSE_COMMIT="xxxx"
-
-# 2024-08-06: including fix waiting for merge:
-CONVERSE_VERSION="livechat_converse_11"
-CONVERSE_COMMIT="1625f9b6ee81fa9ebd8df5cba306bde478f1943b"
-# 2024-08-06: including new getOccupantActionButtons hook (waiting for merge in Converse upstream)
-CONVERSE_COMMIT="5b35bd31c7e26c1f5396cd65eef69293d8b388c1"
-# 2024-08-20: accessibility fixes
-CONVERSE_COMMIT="1e76411e26070709771bf9b099d3c889aeec1878"
 
 rootdir="$(pwd)"
 src_dir="$rootdir/conversejs"